add support for pointpilliars (second)

+++ second/second/data/all_dataset.py
@@ -7,7 +7,6 @@ from second.core import box_np_ops
 from second.data.dataset import Dataset, get_dataset_class
 from second.data.kitti_dataset import KittiDataset
 import second.data.nuscenes_dataset as nuds
-import second.data.my_lyft_dataset as lyds
 from second.utils.progress_bar import progress_bar_iter as prog_bar
 
 from concurrent.futures import ProcessPoolExecutor
@@ -18,11 +17,12 @@ def create_groundtruth_database(dataset_class_name,
                                 used_classes=None,
                                 database_save_path=None,
                                 db_info_save_path=None,
-                                relative_path=False,
+                                relative_path=True,
                                 add_rgb=False,
                                 lidar_only=False,
                                 bev_only=False,
                                 coors_range=None):
+    #import pdb; pdb.set_trace()
     dataset = get_dataset_class(dataset_class_name)(
         info_path=info_path,
         root_path=data_path,
@@ -38,15 +38,11 @@ def create_groundtruth_database(dataset_class_name,
     all_db_infos = {}
 
     group_counter = 0
-    # for j in prog_bar(list(range(len(dataset)))):
-    for j in range(len(dataset)):
+    for j in prog_bar(list(range(len(dataset)))):
         image_idx = j
         sensor_data = dataset.get_sensor_data(j)
         if "image_idx" in sensor_data["metadata"]:
-            if dataset_class_name == "MyLeftDataset":
-                image_idx = sensor_data["metadata"]["token"]
-            else:
-                image_idx = sensor_data["metadata"]["image_idx"]
+            image_idx = sensor_data["metadata"]["image_idx"]
         points = sensor_data["lidar"]["points"]
         annos = sensor_data["lidar"]["annotations"]
         gt_boxes = annos["boxes"]
@@ -62,16 +58,19 @@ def create_groundtruth_database(dataset_class_name,
             difficulty = annos["difficulty"]
 
         num_obj = gt_boxes.shape[0]
-        print(gt_boxes)
         point_indices = box_np_ops.points_in_rbbox(points, gt_boxes)
         for i in range(num_obj):
             filename = f"{image_idx}_{names[i]}_{i}.bin"
             filepath = database_save_path / filename
+
             gt_points = points[point_indices[:, i]]
 
             gt_points[:, :3] -= gt_boxes[i, :3]
-            with open(filepath, 'w') as f:
-                gt_points.tofile(f)
+
+            if not filepath.exists():
+                with open(filepath, 'w') as f:
+                    gt_points.tofile(f)
+
             if (used_classes is None) or names[i] in used_classes:
                 if relative_path:
                     db_path = str(database_save_path.stem + "/" + filename)
@@ -117,7 +116,6 @@ def create_groundtruth_database_parallel(dataset_class_name,
                                 lidar_only=False,
                                 bev_only=False,
                                 coors_range=None):
-
     dataset = get_dataset_class(dataset_class_name)(
         info_path=info_path,
         root_path=data_path,
@@ -132,13 +130,9 @@ def create_groundtruth_database_parallel(dataset_class_name,
     database_save_path.mkdir(parents=True, exist_ok=True)
     all_db_infos = {}
 
-    print(root_path)
-    
-    # for j in prog_bar(list(range(len(dataset)))):
-    for j in range(len(dataset)):
+    for j in prog_bar(list(range(len(dataset)))):
         image_idx = j
         sensor_data = dataset.get_sensor_data(j)
-        print(sensor_data)
         if "image_idx" in sensor_data["metadata"]:
             image_idx = sensor_data["metadata"]["image_idx"]
         points = sensor_data["lidar"]["points"]
